Move useEditPreviewBlock hook to zui
