+++ src/features/surveys/components/SurveyEditor/blocks/ChoiceQuestionBlock.tsx
@@ -23,7 +23,7 @@ import DropdownIcon from 'zui/icons/DropDown';
 import messageIds from 'features/surveys/l10n/messageIds';
 import PreviewableSurveyInput from '../elements/PreviewableSurveyInput';
 import SurveyDataModel from 'features/surveys/models/SurveyDataModel';
-import useEditPreviewBlock from './useEditPreviewBlock';
+import useEditPreviewBlock from 'zui/hooks/useEditPreviewBlock';
 import { ZetkinSurveyOptionsQuestionElement } from 'utils/types/zetkin';
 import { ZUIConfirmDialogContext } from 'zui/ZUIConfirmDialogProvider';
 import ZUIPreviewableInput from 'zui/ZUIPreviewableInput';
